requirements.txt
absl-py==0.12.0
aio-pika==6.8.0
aiofiles==0.6.0
aiohttp==3.6.3
aiormq==3.3.1
APScheduler==3.7.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bidict==0.21.2
boto3==1.17.35
botocore==1.20.35
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorclass==2.2.0
coloredlogs==14.3
colorhash==1.0.3
cryptography==3.4.6
cycler==0.10.0
decorator==4.4.2
dm-tree==0.1.5
dnspython==1.16.0
docopt==0.6.2
fbmessenger==6.0.0
filelock==3.0.12
future==0.18.2
fuzzysearch==0.7.3
fuzzywuzzy==0.18.0
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.36.1
h11==0.9.0
h5py==2.10.0
httpcore==0.11.1
httplib2==0.19.0
httptools==0.1.1
httpx==0.15.4
humanfriendly==9.1
idna==2.10
importlib-metadata==3.7.3
jmespath==0.10.0
joblib==0.15.1
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==2.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
mattermostwrapper==2.2
multidict==4.7.6
networkx==2.5
num2words==0.5.10
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
Pillow==8.1.2
prompt-toolkit==2.0.10
protobuf==3.15.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
PyJWT==2.0.1
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
pyTelegramBotAPI==3.7.6
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-engineio==3.13.2
python-Levenshtein==0.12.2
python-socketio==5.0.0
pytz==2020.5
questionary==1.5.2
rasa==2.4.1
rasa-sdk==2.4.1
redis==3.5.3
regex==2020.9.27
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rocketchat-API==1.15.0
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
s3transfer==0.3.6
sacremoses==0.0.43
sanic==20.9.0
Sanic-Cors==0.10.0.post3
sanic-jwt==1.5.0
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.24.1
scipy==1.6.1
sentry-sdk==0.19.5
six==1.15.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.3.23
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.2
tensorflow-addons==0.12.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
tensorflow-probability==0.11.1
tensorflow-text==2.3.0
termcolor==1.1.0
terminaltables==3.1.0
threadpoolctl==2.1.0
tokenizers==0.10.1
tqdm==4.59.0
twilio==6.50.1
typeguard==2.11.1
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
uvloop==0.14.0
wcwidth==0.2.5
webexteamssdk==1.6
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.1
zipp==3.4.1
